requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
datasets==4.5.0
dill==0.4.0
eval_type_backport==0.3.1
exceptiongroup==1.3.1
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.3.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
iniconfig==2.1.0
joblib==1.5.3
llvmlite==0.43.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.7.1
multiprocess==0.70.18
nltk==3.9.2
numba==0.60.0
numpy==2.0.2
packaging==26.0
pandas==2.3.3
platformdirs==4.4.0
pluggy==1.6.0
propcache==0.4.1
protobuf==6.33.5
pyarrow==21.0.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.3
sentry-sdk==2.53.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
tomli==2.4.0
tqdm==4.67.3
typer==0.23.2
typer-slim==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wandb==0.25.0
xxhash==3.6.0
yarl==1.22.0
